chore: Move business logic out of TradeController into service layer (#1259)
